implement filtering for sun, moon, and rising signs. clean up comments and remove console logs
