+++ frontend/src/components/Chat/ChatSidebar/Sidebar.js
@@ -1,16 +1,14 @@
 import { useContext, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import {socket, ChatContext} from '../../../context/chatContext';
-import io from 'socket.io-client'
 import './Sidebar.css'
 
 function Sidebar () {
   const user = useSelector(state => state.session.user);
 
-  const { setMembers, members, room, currentRoom, setCurrentRoom } = useContext(ChatContext)
+  const { setMembers, members, currentRoom, setCurrentRoom } = useContext(ChatContext)
 
-  const userId = user ? user._id : 1;
-  const [socketConnected, setSocketConnected] = useState(false);
+  // const [socketConnected, setSocketConnected] = useState(false);
 
   const [storeRoom, setStoreRoom] = useState('')
 
@@ -19,11 +17,10 @@ function Sidebar () {
     socket.off("new-user").on("new-user", (users) => {
       setCurrentRoom(storeRoom)
       socket.emit('join-room', currentRoom)
-      console.log(users)
       setMembers(users)
     })
-    socket.on("connection", () => setSocketConnected(true))
-  }, [])
+    // socket.on("connection", () => setSocketConnected(true))
+  }, [storeRoom, currentRoom, setCurrentRoom, setMembers])
 
   const joinRoom = (e, isPublic = true) => {
     const room =String(e.target.id)
@@ -37,7 +34,9 @@ function Sidebar () {
     return <li key={i}
     id={member._id}
     onClick={joinRoom}>{member.name}</li>
-  }}
+    }
+    return null;
+  }
   )
 
   return (
