+++ frontend/src/components/Chat/MessageForm/MessageForm.js
@@ -1,6 +1,6 @@
 import './MessageForm.css'
 import { useSelector } from 'react-redux'
-import { useState, useContext, useEffect } from 'react'
+import { useState, useContext } from 'react'
 import { ChatContext } from '../../../context/chatContext'
 
 function MessageForm () {
@@ -9,15 +9,10 @@ function MessageForm () {
   const user = useSelector(state => state.session.user)
   const userName = useSelector(state => state.session.user.name)
 
-  // console.log(user)
   const [msg, setMsg] = useState('');
   const { socket, messages, setMessages, currentRoom } = useContext(ChatContext);
 
   socket.off('room-messages').on('room-messages', (roomMessages) => {
-    console.log("room messages!")
-    console.log(roomMessages[0])
-    console.log(`user is: ${userName}`)
-    console.log("room messages!")
     setMessages(roomMessages[0])
   })
 
@@ -30,12 +25,11 @@ function MessageForm () {
     setMsg("")
   }
 
-  const [dynamicMessagesList, setDynamicMessagesList] = useState('')
+  // const [dynamicMessagesList, setDynamicMessagesList] = useState('')
 
   let messagesList = messages?.messagesByDate ? Object.values(messages.messagesByDate) : null;
   messagesList = messagesList?.map((message, i) => <li key={i} id={message._id} className="chat-message">{message.content}</li>)
 
-  // const messagesList = Object.values(messages?.messagesByDate).map((message, i) => <li key={i} id={message._id} className="chat-message">{message.content}</li>).reverse()
 
   return (
     <>
@@ -45,8 +39,7 @@ function MessageForm () {
         {messagesList}
       </div>
       <h1>from: {userName}</h1>
-      {/* </div> */}
-      {/* </div> */}
+
 
       <form onSubmit={handleSubmit}
         className="message-form">
