+++ frontend/src/components/Discover/Discover.css
@@ -38,14 +38,6 @@
     flex-wrap: wrap;
 }
 
-.page-filter {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    width: 35%;
-    margin-bottom: 2%;
-    font-size: 16px;
-}
 
 .heading {
     font-size: 3rem;
@@ -60,4 +52,26 @@ select {
     padding: 0 1% 0 1%;
     line-height: 0;
     width: 60px;
+    margin: 0 25px 0 5px;
+}
+#filters button {
+    background-color: #5b0e7a;
+    height: 15px;
+    margin: 0 15px 0 5px;
+    font-size: 12px;
+    font-weight: bold;
+    color: #d3a02099;
+    border: 0;
+    border-radius: 4px;
+    transition: 0.2s;
+}
+#filters {
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    overflow-x: hidden;
+    width: 1000px;
+    margin-bottom: 20px;
+    margin-left: 20px;
+    font-size: 2.1vh;
 }
