+++ frontend/src/components/Matches/Matches.css
@@ -1,49 +1,42 @@
-.matches-container {
-    display: flex;
-    flex-wrap: wrap;
+.matches-outer-container {
     background-color: #510e6f;
-    min-height: 90vh;
-    height: 100%;
-    width: 100%;
+    min-height: 100vh;
     padding-bottom: 50px;
-    margin-top: -4%;
     padding-top: 2%;
     padding-left: 3%;
     color: #d3a02099;
     font-family:'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
 }
 
+.matches-container {
+    display: flex;
+    margin: 0 2% 0 1%;
+}
+
 
 .heading {
     font-size: 3rem;
-    margin-bottom: 0;
     font-weight: bold;
-    margin: 3% 0 0 0;
+    margin: 0;
 }
 
-.page-filter {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    width: 35%;
-    margin-bottom: 2%;
-    font-size: 16px;
-    /* margin-right: 65%; */
-}
-
-/* .matches-container h2{
-    min-width: 100%;
-} */
 
 .matches-right{
     display: flex;
     width: 40%;
-    border: 1px solid red;
-    /* background-color: mediumpurple; */
+    border-radius: 8px;
+    box-shadow: 0 0 3px 3px #00000050;
+    width: 60%;
+    height: 500px;
+    padding: 20px;
 }
 
-.matches-left{
+.matches-left {
     display: flex;
     flex-direction: column;
     width: 50%;
+}
+
+.matches-left .user-card {
+    width: 70%
 }
