+++ frontend/src/components/Profile/Profile.css
@@ -7,7 +7,7 @@
    bottom: 15%;
    width: 30vw;
    height: 30vw;
-  background: radial-gradient(circle, rgba(113,14,143,1) 0%, rgba(74,23,84,1) 35%, rgba(14,0,20,1) 100%);
+   background: radial-gradient(circle, rgba(113,14,143,1) 0%, rgba(74,23,84,1) 35%, rgba(14,0,20,1) 100%);
    border-top: .8vw ridge rgba(233, 193, 93);
    border-left: .8vw ridge rgba(233, 193, 93);
    border-bottom: .8vw groove rgba(233, 193, 93);
